No Ticket: Switch to the Ibotta AspectJ Plugin
feat : The plugin basically seems to switch right in without
any issues directly. Since both the old android-gradle-aspectj
plugin and this ibotta plugin are both using the same aspectjrt
classes, no code change is required other than in the gradle
files.

The reason for the switch on my end is that the android-gradle-aspectj
project is slow to make updates and doesn't seem to be
frequently maintained these days. the ibotta plugin doesn't
seem to break as easily between AGP and Gradle updates, but
I am reaching out to the devs in order to confirm.